+++ OBAKit/Models/unmanaged/OBAStopV2.h
@@ -1,9 +1,10 @@
 #import "OBAHasReferencesV2.h"
 @import MapKit;
+#import "OBARouteType.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface OBAStopV2 : OBAHasReferencesV2 <MKAnnotation>
+@interface OBAStopV2 : OBAHasReferencesV2 <MKAnnotation,NSCoding,NSCopying>
 
 @property (nonatomic, strong) NSString * stopId;
 @property (nonatomic, strong) NSString * name;
@@ -11,17 +12,19 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, strong) NSString * direction;
 @property (nonatomic, strong) NSNumber * latitude;
 @property (nonatomic, strong) NSNumber * longitude;
-@property (nonatomic, strong) NSArray * routeIds;
+@property (nonatomic, strong) NSArray<NSString*> *routeIds;
 
-@property (weak, nonatomic,readonly) NSArray * routes;
+@property(nonatomic,strong,readonly) NSArray<OBARouteV2*> *routes;
 
 @property (nonatomic,readonly) double lat;
 @property (nonatomic,readonly) double lon;
 @property (nonatomic,readonly) CLLocationCoordinate2D coordinate;
 
-- (NSComparisonResult) compareUsingName:(OBAStopV2*)aStop;
+- (NSComparisonResult)compareUsingName:(OBAStopV2*)aStop;
 
-- (NSString*) routeNamesAsString;
+- (NSString*)routeNamesAsString;
+
+- (OBARouteType)firstAvailableRouteTypeForStop;
 
 @end
