+++ org.onebusaway.iphone.xcodeproj/xcshareddata/xcschemes/OneBusAway.xcscheme
@@ -75,6 +75,10 @@
       </BuildableProductRunnable>
       <AdditionalOptions>
       </AdditionalOptions>
+      <LocationScenarioReference
+         identifier = "../gpx_files/capitolhill.gpx"
+         referenceType = "0">
+      </LocationScenarioReference>
    </LaunchAction>
    <ProfileAction
       displayScaleIsEnabled = "NO"
